requirements.txt
# Core dependencies
python-telegram-bot>=20.3
flask>=2.3.3
flask-cors>=4.0.0
requests>=2.31.0
numpy>=1.24.0
python-dotenv>=1.0.0
openai>=1.0.0

# PDF Processing
pypdf>=3.17.2
pdfminer.six>=20221105
pdf2image>=1.16.3
pytesseract>=0.3.10
pymupdf>=1.23.8

# NLP and machine learning
nltk>=3.8.1
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
langdetect>=1.0.9

# Vector Database
qdrant-client>=1.6.0
pinecone-client>=2.2.2
weaviate-client>=3.24.0

# Speech recognition and synthesis
SpeechRecognition>=3.10.0
elevenlabs>=0.2.24
pydub>=0.25.1

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Logging and monitoring
rich>=13.5.2
loguru>=0.7.0
prometheus-client>=0.17.1
sentry-sdk>=1.32.0

# Development and testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.1

# Caching and performance
redis>=5.0.0
cachetools>=5.3.1

# Utilities
tqdm>=4.66.1
tenacity>=8.2.3

# Security
pyjwt>=2.3.0
cryptography>=38.0.1
bcrypt>=4.0.0

# Database integrations
pymysql>=1.1.0

# Machine learning and AI
tensorflow>=2.14.0
torch>=2.0.1
transformers>=4.31.0
keras>=2.14.0

# Data processing and visualization
pandas>=2.0.3
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.16.1

# Additional utilities
python-dotenv>=0.20.0
pydantic>=1.9.1
ujson>=5.4.0

# Database requirements
SQLAlchemy>=1.4.40

# Additional data sources
kaggle>=1.5.16
boto3>=1.28.38
google-api-python-client>=2.97.0

# Dataset Connectors
huggingface_hub>=0.17.1
datasets>=2.14.0
xmltodict>=0.13.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Blockchain integration
toncenter>=0.0.4
pytonlib>=0.0.20
tonweb>=0.0.34

# AR/VR support
pypng>=0.20220715.0
qrcode>=7.4.2
threejs-jupyter>=2.10.0

# Multimodal and gesture
mediapipe>=0.10.0
imgaug>=0.4.0

# Documentation
sphinx>=7.2.5
sphinx-rtd-theme>=1.3.0

# Additional AI capabilities
sympy>=1.12.0
networkx>=3.1
gym>=0.26.0
stable-baselines3>=2.1.0

# Additional data sources
snowflake-connector-python>=3.0.4
azure-cosmos>=4.5.1
ibm_db>=3.1.5
ibm_db_dbi>=3.1.4

# Additional utilities
pydantic>=1.9.1
ujson>=5.4.0

# Database requirements
SQLAlchemy>=1.4.40

# JWT Authentication
PyJWT>=2.4.0
cryptography>=37.0.0 